Consistent use of "Import File" throughout editors
Replaced instances of "Add File" with "Import File".

Added tooltip to "Import File" menu item.

Changed "Import <type> (<count>)" button text to
"Import as <type> (<count>)" to emphasize that we already have a subject
(the file) in mind.  Without the "as", "Import image" and "Import File"
would seem to do the same thing.

Changed "Asset Pack File" Property section title to "Asset Pack Entry".
The other Property sections titles for e.g. a PNG refer to aspects of
*that PNG*: "File" (details when thinking of it as a file), "Image"
(details when thinking of it as an image), "Asset Pack Entry" (details
when thinking of it as an Asset Pack entry).  Otherwise, "Asset Pack
File" makes it seem like the selected PNG could be treated as an Asset
Pack File.

(I was going to call that Property section "Import File", but then I
noticed that doesn't make sense when selecting an already-imported file
in the Files view.  In that case the Property section points the user to
the existing Asset Pack entry.)